push 0.0.16 to git
