+++ .env.example
@@ -1,0 +1,7 @@
+NODE_ENV=production
+PORT=3000
+GH_TOKEN=your_github_token
+API_TOKEN=your_api_token
+API_IP=""
+API_IP_REGEX=""
+KENER_BASE_PATH=""
