+++ database/monitors.json
@@ -1,0 +1,54 @@
+[
+    {
+        "name": "OkBookmarksX",
+        "description": "Stop forgetting about your bookmarks <a href=\"https://okbookmarks.com/\" class=\"font-medium underline underline-offset-4\" target=\"_blank\">https://okbookmarks.com/</a>",
+        "tag": "okbookmarks",
+        "image": "https://okbookmarks.com/app/mybookmark.png",
+        "api": {
+            "method": "GET",
+            "url": "https://okbookmarks.com/",
+            "eval": "(function (statusCode, responseTime, responseData) {\n\tlet statusCodeShort = Math.floor(statusCode/100);\n    if(statusCode == 429 || (statusCodeShort >=2 && statusCodeShort <= 3)) {\n        return {\n\t\t\tstatus: 'UP',\n\t\t\tlatency: responseTime,\n        }\n    } \n\treturn {\n\t\tstatus: 'DOWN',\n\t\tlatency: responseTime,\n\t}\n})",
+            "timeout": 10000
+        },
+        "folderName": "okbookmarksx",
+        "dayDegradedMinimumCount": 1,
+        "dayDownMinimumCount": 1,
+        "includeDegradedInDowntime": false,
+        "path0Day": "./database/okbookmarksx.0day.utc.json",
+        "path90Day": "./database/okbookmarksx.90day.utc.json",
+        "hasAPI": true
+    },
+    {
+        "name": "Google Search",
+        "description": "Search the world's information, including webpages, images, videos and more. <a href=\"https://www.google.com/webhp\" class=\"font-medium underline underline-offset-4\" target=\"_blank\">https://www.google.com/webhp</a>",
+        "tag": "google-search",
+        "image": "/google.png",
+        "api": {
+            "method": "GET",
+            "url": "https://www.google.com/webhp",
+            "eval": "(function (statusCode, responseTime, responseData) {\n\tlet statusCodeShort = Math.floor(statusCode/100);\n    if(statusCode == 429 || (statusCodeShort >=2 && statusCodeShort <= 3)) {\n        return {\n\t\t\tstatus: 'UP',\n\t\t\tlatency: responseTime,\n        }\n    } \n\treturn {\n\t\tstatus: 'DOWN',\n\t\tlatency: responseTime,\n\t}\n})",
+            "timeout": 10000
+        },
+        "folderName": "google-search",
+        "dayDegradedMinimumCount": 1,
+        "dayDownMinimumCount": 1,
+        "includeDegradedInDowntime": false,
+        "path0Day": "./database/google-search.0day.utc.json",
+        "path90Day": "./database/google-search.90day.utc.json",
+        "hasAPI": true
+    },
+    {
+        "name": "Earth",
+        "description": "Our blue planet",
+        "tag": "earth",
+        "defaultStatus": "UP",
+        "image": "/earth.png",
+        "folderName": "earth",
+        "dayDegradedMinimumCount": 1,
+        "dayDownMinimumCount": 1,
+        "includeDegradedInDowntime": false,
+        "path0Day": "./database/earth.0day.utc.json",
+        "path90Day": "./database/earth.90day.utc.json",
+        "hasAPI": false
+    }
+]
