+++ database/okbookmarks.90day.utc.json
@@ -1,0 +1,1 @@
+{"_90Day":{"1723507200":{"timestamp":1723507200,"cssClass":"api-nodata","message":"No Data"},"1723593600":{"timestamp":1723593600,"cssClass":"api-nodata","message":"No Data"},"1723680000":{"timestamp":1723680000,"cssClass":"api-nodata","message":"No Data"},"1723766400":{"timestamp":1723766400,"cssClass":"api-nodata","message":"No Data"},"1723852800":{"timestamp":1723852800,"cssClass":"api-nodata","message":"No Data"},"1723939200":{"timestamp":1723939200,"cssClass":"api-nodata","message":"No Data"},"1724025600":{"timestamp":1724025600,"cssClass":"api-nodata","message":"No Data"},"1724112000":{"timestamp":1724112000,"cssClass":"api-nodata","message":"No Data"},"1724198400":{"timestamp":1724198400,"cssClass":"api-nodata","message":"No Data"},"1724284800":{"timestamp":1724284800,"cssClass":"api-nodata","message":"No Data"},"1724371200":{"timestamp":1724371200,"cssClass":"api-nodata","message":"No Data"},"1724457600":{"timestamp":1724457600,"cssClass":"api-nodata","message":"No Data"},"1724544000":{"timestamp":1724544000,"cssClass":"api-nodata","message":"No Data"},"1724630400":{"timestamp":1724630400,"cssClass":"api-nodata","message":"No Data"},"1724716800":{"timestamp":1724716800,"cssClass":"api-nodata","message":"No Data"},"1724803200":{"timestamp":1724803200,"cssClass":"api-nodata","message":"No Data"},"1724889600":{"timestamp":1724889600,"cssClass":"api-nodata","message":"No Data"},"1724976000":{"timestamp":1724976000,"cssClass":"api-nodata","message":"No Data"},"1725062400":{"timestamp":1725062400,"cssClass":"api-nodata","message":"No Data"},"1725148800":{"timestamp":1725148800,"cssClass":"api-nodata","message":"No Data"},"1725235200":{"timestamp":1725235200,"cssClass":"api-nodata","message":"No Data"},"1725321600":{"timestamp":1725321600,"cssClass":"api-nodata","message":"No Data"},"1725408000":{"timestamp":1725408000,"cssClass":"api-nodata","message":"No Data"},"1725494400":{"timestamp":1725494400,"cssClass":"api-nodata","message":"No Data"},"1725580800":{"timestamp":1725580800,"cssClass":"api-nodata","message":"No Data"},"1725667200":{"timestamp":1725667200,"cssClass":"api-nodata","message":"No Data"},"1725753600":{"timestamp":1725753600,"cssClass":"api-nodata","message":"No Data"},"1725840000":{"timestamp":1725840000,"cssClass":"api-nodata","message":"No Data"},"1725926400":{"timestamp":1725926400,"cssClass":"api-nodata","message":"No Data"},"1726012800":{"timestamp":1726012800,"cssClass":"api-nodata","message":"No Data"},"1726099200":{"timestamp":1726099200,"cssClass":"api-nodata","message":"No Data"},"1726185600":{"timestamp":1726185600,"cssClass":"api-nodata","message":"No Data"},"1726272000":{"timestamp":1726272000,"cssClass":"api-nodata","message":"No Data"},"1726358400":{"timestamp":1726358400,"cssClass":"api-nodata","message":"No Data"},"1726444800":{"timestamp":1726444800,"cssClass":"api-nodata","message":"No Data"},"1726531200":{"timestamp":1726531200,"cssClass":"api-nodata","message":"No Data"},"1726617600":{"timestamp":1726617600,"cssClass":"api-nodata","message":"No Data"},"1726704000":{"timestamp":1726704000,"cssClass":"api-nodata","message":"No Data"},"1726790400":{"timestamp":1726790400,"cssClass":"api-nodata","message":"No Data"},"1726876800":{"timestamp":1726876800,"cssClass":"api-nodata","message":"No Data"},"1726963200":{"timestamp":1726963200,"cssClass":"api-nodata","message":"No Data"},"1727049600":{"timestamp":1727049600,"cssClass":"api-nodata","message":"No Data"},"1727136000":{"timestamp":1727136000,"cssClass":"api-nodata","message":"No Data"},"1727222400":{"timestamp":1727222400,"cssClass":"api-nodata","message":"No Data"},"1727308800":{"timestamp":1727308800,"cssClass":"api-nodata","message":"No Data"},"1727395200":{"timestamp":1727395200,"cssClass":"api-nodata","message":"No Data"},"1727481600":{"timestamp":1727481600,"cssClass":"api-nodata","message":"No Data"},"1727568000":{"timestamp":1727568000,"cssClass":"api-nodata","message":"No Data"},"1727654400":{"timestamp":1727654400,"cssClass":"api-nodata","message":"No Data"},"1727740800":{"timestamp":1727740800,"cssClass":"api-nodata","message":"No Data"},"1727827200":{"timestamp":1727827200,"cssClass":"api-nodata","message":"No Data"},"1727913600":{"timestamp":1727913600,"cssClass":"api-nodata","message":"No Data"},"1728000000":{"timestamp":1728000000,"cssClass":"api-nodata","message":"No Data"},"1728086400":{"timestamp":1728086400,"cssClass":"api-nodata","message":"No Data"},"1728172800":{"timestamp":1728172800,"cssClass":"api-nodata","message":"No Data"},"1728259200":{"timestamp":1728259200,"cssClass":"api-nodata","message":"No Data"},"1728345600":{"timestamp":1728345600,"cssClass":"api-nodata","message":"No Data"},"1728432000":{"timestamp":1728432000,"cssClass":"api-nodata","message":"No Data"},"1728518400":{"timestamp":1728518400,"cssClass":"api-nodata","message":"No Data"},"1728604800":{"timestamp":1728604800,"cssClass":"api-nodata","message":"No Data"},"1728691200":{"timestamp":1728691200,"cssClass":"api-nodata","message":"No Data"},"1728777600":{"timestamp":1728777600,"cssClass":"api-nodata","message":"No Data"},"1728864000":{"timestamp":1728864000,"cssClass":"api-nodata","message":"No Data"},"1728950400":{"timestamp":1728950400,"cssClass":"api-nodata","message":"No Data"},"1729036800":{"timestamp":1729036800,"cssClass":"api-nodata","message":"No Data"},"1729123200":{"timestamp":1729123200,"cssClass":"api-nodata","message":"No Data"},"1729209600":{"timestamp":1729209600,"cssClass":"api-nodata","message":"No Data"},"1729296000":{"timestamp":1729296000,"cssClass":"api-nodata","message":"No Data"},"1729382400":{"timestamp":1729382400,"cssClass":"api-nodata","message":"No Data"},"1729468800":{"timestamp":1729468800,"cssClass":"api-nodata","message":"No Data"},"1729555200":{"timestamp":1729555200,"cssClass":"api-nodata","message":"No Data"},"1729641600":{"timestamp":1729641600,"cssClass":"api-nodata","message":"No Data"},"1729728000":{"timestamp":1729728000,"cssClass":"api-nodata","message":"No Data"},"1729814400":{"timestamp":1729814400,"cssClass":"api-nodata","message":"No Data"},"1729900800":{"timestamp":1729900800,"cssClass":"api-nodata","message":"No Data"},"1729987200":{"timestamp":1729987200,"cssClass":"api-nodata","message":"No Data"},"1730073600":{"timestamp":1730073600,"cssClass":"api-nodata","message":"No Data"},"1730160000":{"timestamp":1730160000,"cssClass":"api-nodata","message":"No Data"},"1730246400":{"timestamp":1730246400,"cssClass":"api-nodata","message":"No Data"},"1730332800":{"timestamp":1730332800,"cssClass":"api-nodata","message":"No Data"},"1730419200":{"timestamp":1730419200,"cssClass":"api-nodata","message":"No Data"},"1730505600":{"timestamp":1730505600,"cssClass":"api-nodata","message":"No Data"},"1730592000":{"timestamp":1730592000,"cssClass":"api-nodata","message":"No Data"},"1730678400":{"timestamp":1730678400,"cssClass":"api-nodata","message":"No Data"},"1730764800":{"timestamp":1730764800,"cssClass":"api-nodata","message":"No Data"},"1730851200":{"timestamp":1730851200,"cssClass":"api-nodata","message":"No Data"},"1730937600":{"timestamp":1730937600,"cssClass":"api-nodata","message":"No Data"},"1731024000":{"timestamp":1731024000,"cssClass":"api-nodata","message":"No Data"},"1731110400":{"timestamp":1731110400,"cssClass":"api-nodata","message":"No Data"},"1731196800":{"timestamp":1731196800,"cssClass":"api-nodata","message":"No Data"},"1731283200":{"timestamp":1731283200,"cssClass":"api-up","message":"Status OK"}},"uptime0Day":"100","uptime90Day":"100","dailyUps":16,"dailyDown":0,"dailyDegraded":0}
