--- dev.js
@@ -1,2 +1,0 @@
-import { Startup } from "./scripts/startup.js";
-Startup();
