--- docker/root/etc/s6-overlay/s6-rc.d/init-app-config/type
@@ -1,1 +1,0 @@
-oneshot
