--- docker/root/etc/s6-overlay/s6-rc.d/init-app-config/up
@@ -1,1 +1,0 @@
-/etc/s6-overlay/s6-rc.d/init-app-config/run
