--- docker/root/etc/s6-overlay/s6-rc.d/svc-app/run
@@ -1,7 +1,0 @@
-#!/usr/bin/with-contenv bash
-
-echo -e "\nApp is starting!"
-export NODE_ENV=production
-cd /app || exit
-exec \
-	  s6-setuidgid abc /usr/bin/node $NODE_ARGS /app/prod.js
