--- docker/root/etc/s6-overlay/s6-rc.d/svc-app/type
@@ -1,1 +1,0 @@
-longrun
