+++ main.js
@@ -1,6 +1,9 @@
 import { handler } from "./build/handler.js";
+import dotenv from "dotenv";
+dotenv.config();
 import express from "express";
 const PORT = process.env.PORT || 3000;
+
 const app = express();
 app.use(handler);
