+++ src/lib/locales/hi.json
@@ -5,7 +5,7 @@
 		"resolved": "रेसोल्वेड"
 	},
 	"monitor": {
-		"90_day": "९० दिन",
+		"90_day": "90 दिन",
 		"am": "ऍम",
 		"badge": "बैज",
 		"badge_desc": "इस मॉनिटर के लिए SVG बैज प्राप्त करें",
