+++ src/routes/(kener)/+page.server.js
@@ -2,11 +2,12 @@
 import { Mapper, GetOpenIncidents, FilterAndInsertMonitorInIncident } from "$lib/server/github.js";
 import { FetchData } from "$lib/server/page";
 import fs from "fs-extra";
-
-import monitorJSON from "$lib/server/data/monitors.json?raw";
+import { monitorsStore } from "$lib/server/stores/monitors";
+import { get } from "svelte/store";
 
 export async function load({ parent }) {
-	let monitors = JSON.parse(monitorJSON);
+	let monitors = get(monitorsStore);
+
 	const parentData = await parent();
 	const siteData = parentData.site;
 	const github = siteData.github;
