+++ src/routes/(kener)/api/incident/[incidentNumber]/status/+server.js
@@ -3,7 +3,9 @@
 import { json } from "@sveltejs/kit";
 import { auth, GHIssueToKenerIncident } from "$lib/server/webhook";
 import { UpdateIssueLabels, GetIncidentByNumber } from "$lib/server/github";
-import siteJSON from "$lib/server/data/site.json?raw";
+import { siteStore } from "$lib/server/stores/site";
+import { get } from "svelte/store";
+
 import fs from "fs-extra";
 
 export async function POST({ request, params }) {
@@ -43,7 +45,7 @@ export async function POST({ request, params }) {
 		);
 	}
 
-	let site = JSON.parse(siteJSON);
+	let site = get(siteStore);
 	let github = site.github;
 
 	let issue = await GetIncidentByNumber(github, incidentNumber);
