+++ src/routes/(kener)/monitor-[tag]/+page.server.js
@@ -1,11 +1,12 @@
 // @ts-nocheck
 import { Mapper, GetOpenIncidents, FilterAndInsertMonitorInIncident } from "$lib/server/github.js";
 import { FetchData } from "$lib/server/page";
-import monitorJSON from "$lib/server/data/monitors.json?raw";
-import fs from "fs-extra";
+import { monitorsStore } from "$lib/server/stores/monitors";
+import { get } from "svelte/store";
 
 export async function load({ params, route, url, parent }) {
-	let monitors = JSON.parse(monitorJSON);
+	let monitors = get(monitorsStore);
+
 	const parentData = await parent();
 	const siteData = parentData.site;
 	const github = siteData.github;
